+++ student_self_organization-service/src/main/java/de/fherfurt/organization/models/Answer.java
@@ -1,4 +1,4 @@
-package de.fherfurt.organization.forum.core;
+package de.fherfurt.organization.models;
 
 import de.fherfurt.organization.storage.core.AbstractDatabaseEntity;
