+++ student_self_organization-service/src/main/java/de/fherfurt/organization/resources/QuestionResource.java
@@ -1,6 +1,6 @@
 package de.fherfurt.organization.resources;
 
-import de.fherfurt.organization.forum.core.Question;
+import de.fherfurt.organization.models.Question;
 import de.fherfurt.organization.storage.core.RepositoryFactory;
 import de.fherfurt.organization.storage.repository.QuestionRepository;
 import jakarta.ws.rs.*;
