+++ student_self_organization-service/src/main/java/de/fherfurt/organization/storage/core/IFaqDao.java
@@ -1,13 +1,11 @@
-package de.fherfurt.organization.faq.core;
+package de.fherfurt.organization.storage.core;
 
-import de.fherfurt.organization.faq.model.Element;
+import de.fherfurt.organization.models.Element;
 import de.fherfurt.organization.storage.core.IGenericDao;
 
 import java.util.List;
 
 public interface IFaqDao extends IGenericDao<Element> {
-
-    Element findElementById(int id);
 
     List<Element> findElementByTitle(String title);
