+++ student_self_organization-service/src/main/java/de/fherfurt/organization/todo/repository/TodoRepository.java
@@ -1,6 +1,6 @@
 package de.fherfurt.organization.todo.repository;
 
-import de.fherfurt.organization.faq.core.errors.EntryNotFoundException;
+import de.fherfurt.organization.errors.EntryNotFoundException;
 import de.fherfurt.organization.todo.core.Task;
 
 import java.util.LinkedList;
